+++ pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.github.a-pz</groupId>
 	<artifactId>struts2-thymeleaf3-plugin</artifactId>
-	<version>6.0.0-RELEASE</version>
+	<version>7.0.0-RELEASE</version>
 	<packaging>jar</packaging>
 	<name>Struts2-Thymeleaf3-plugin</name>
 	<url>https://github.com/A-pZ/struts2-thymeleaf3-plugin</url>
@@ -41,19 +41,14 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<struts2.version>6.0.0</struts2.version>
-		<thymeleaf.version>3.0.12.RELEASE</thymeleaf.version>
-		<spring.version>5.3.22</spring.version>
+		<struts2.version>7.0.0</struts2.version>
+		<thymeleaf.version>3.1.2.RELEASE</thymeleaf.version>
+		<spring.version>6.2.1</spring.version>
 	</properties>
 	<repositories>
 		<repository>
 			<id>maven-org-repo</id>
-			<url>http://repo1.maven.org/maven2/</url>
-		</repository>
-		<repository>
-			<id>apache.snapshots</id>
-			<name>ASF Maven 2 Snapshot</name>
-			<url>https://repository.apache.org/content/groups/snapshots/</url>
+			<url>https://repo1.maven.org/maven2/</url>
 		</repository>
 	</repositories>
 	<distributionManagement>
@@ -78,7 +73,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.thymeleaf</groupId>
-			<artifactId>thymeleaf-spring4</artifactId>
+			<artifactId>thymeleaf-spring5</artifactId>
 			<version>${thymeleaf.version}</version>
 		</dependency>
 		<dependency>
@@ -92,21 +87,20 @@
 			<version>${spring.version}</version>
 		</dependency>
 		<dependency>
-			<groupId>javax.servlet</groupId>
-			<artifactId>javax.servlet-api</artifactId>
-			<version>3.1.0</version>
+			<groupId>org.springframework</groupId>
+			<artifactId>spring-core</artifactId>
+			<version>${spring.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>jakarta.servlet</groupId>
+			<artifactId>jakarta.servlet-api</artifactId>
+			<version>6.1.0</version>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
-			<version>1.18.24</version>
-		</dependency>
-		<!-- https://mvnrepository.com/artifact/org.apache.commons/commons-text -->
-		<dependency>
-		    <groupId>org.apache.commons</groupId>
-		    <artifactId>commons-text</artifactId>
-		    <version>1.9</version>
+			<version>1.18.36</version>
 		</dependency>
 	</dependencies>
 	<build>
@@ -118,8 +112,8 @@
 					<fork>true</fork>
 					<meminitial>256m</meminitial>
 					<maxmem>512m</maxmem>
-					<source>11</source>
-					<target>11</target>
+					<source>17</source>
+					<target>17</target>
 					<encoding>${project.build.sourceEncoding}</encoding>
 				</configuration>
 			</plugin>
@@ -156,8 +150,8 @@
 				<configuration>
 					<downloadJavadocs>true</downloadJavadocs>
 					<downloadSources>true</downloadSources>
-					<jdkName>1.8</jdkName>
-					<jdkLevel>1.8</jdkLevel>
+					<jdkName>17</jdkName>
+					<jdkLevel>17</jdkLevel>
 				</configuration>
 			</plugin>
 		</plugins>
