+++ src/main/java/com/github/apz/struts/plugins/thymeleaf/diarect/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author A-pZ
+ *
+ */
+package com.github.apz.struts.plugins.thymeleaf.diarect;
