+++ src/main/java/com/github/apz/struts/plugins/thymeleaf/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * @author A-pZ
+ *
+ */
+package com.github.apz.struts.plugins.thymeleaf;
