--- src/main/java/serendip/struts/plugins/thymeleaf/diarect/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author A-pZ
- *
- */
-package serendip.struts.plugins.thymeleaf.diarect;
