--- src/main/java/serendip/struts/plugins/thymeleaf/spi/TemplateEngineProvider.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2013 Steven Benitez.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package serendip.struts.plugins.thymeleaf.spi;
-
-import org.thymeleaf.TemplateEngine;
-
-/**
- * Provides Struts with a Thymeleaf {@link org.thymeleaf.TemplateEngine} to use
- * when rendering templates.
- *
- * @author A-pZ ( Original source : Steven Benitez )
- * @see DefaultTemplateEngineProvider
- * @since 2.3.15
- */
-public interface TemplateEngineProvider {
-  /**
-   * Returns the TemplateEngine to use to render a Thymeleaf template. This
-   * method will be called repeatedly, so implementations should create the
-   * actual TemplateEngine object outside of this method. For an example, refer
-   * to {@link DefaultTemplateEngineProvider}.
-   *
-   * @return The TemplateEngine to use to render a Thymeleaf template.
-   */
-  TemplateEngine get();
-}
