--- src/main/java/serendip/struts/plugins/thymeleaf/spring/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author A-pZ
- *
- */
-package serendip.struts.plugins.thymeleaf.spring;
